Add files via upload
This project was my first exam in coding school. We were tasked with reproducing a web page based only on an example image (see pg example.png) and the necessary graphics —and all within 4.5 hours. While I have to admit that it probably would have been faster to recreate the project from scratch than to clean up my beginner's code from the first two weeks of school, it was an enjoyable challenge to perfect the project for my portfolio as a graduate.